Move train module into experimental
