+++ runtime/onert/api/python/package/experimental/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["train"]
+
+from . import train
