+++ runtime/onert/api/python/package/experimental/train/__init__.py
@@ -1,5 +1,5 @@
 from .session import TrainSession as session
-from ..native.libnnfw_api_pybind import traininfo
+from onert.native.libnnfw_api_pybind import traininfo
 from .dataloader import DataLoader
 from . import optimizer
 from . import losses
