+++ runtime/onert/api/python/package/experimental/train/losses/__init__.py
@@ -1,5 +1,5 @@
 from .cce import CategoricalCrossentropy
 from .mse import MeanSquaredError
-from ...native.libnnfw_api_pybind import lossinfo
+from onert.native.libnnfw_api_pybind import lossinfo
 
 __all__ = ["CategoricalCrossentropy", "MeanSquaredError", "lossinfo", "loss"]
