+++ runtime/onert/api/python/package/experimental/train/optimizer/__init__.py
@@ -1,5 +1,5 @@
 from .sgd import SGD
 from .adam import Adam
-from ...native.libnnfw_api_pybind import trainable_ops
+from onert.native.libnnfw_api_pybind import trainable_ops
 
 __all__ = ["SGD", "Adam", "trainable_ops"]
